requirements.txt
Django==2.2.4
django-appconf==1.0.3
django-better-admin-arrayfield==1.3.0
django-compressor==2.3
django-cors-headers==3.1.0
django-elasticsearch-dsl==7.0.0
django-elasticsearch-dsl-drf==0.20.1
django-libsass==0.7
django-modeltranslation==0.13.3
django-nine==0.2.2
djangorestframework==3.10.2
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
libsass==0.19.2
numpy==1.17.1
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
xlrd==1.2.0
